fix(iOS, layout): fix iOS 16 layout
This (partially) reverts commit 9e9e3632605fa12870bab401c4c0cf52e2c37299.

We're reverting this to fix the iOS 16 layout problems caused by the interaction between the app rootViewControoler and JitsiMeetViewController, with the downside that using any RN components based on native view controllers will most likely be faulty. The solution for this would be to change the SDK's entry point to a view controller, instead of JitsiMeetView.